Revert "dd"
This reverts commit 9f0be80f08e0f62cce1348f2ad9f7106aa69e976.

+++ klimatapp/Models/Measurement.cs
@@ -16,7 +16,7 @@ namespace klimatapp.Models
         /// <summary>
         /// Type of measurement
         /// </summary>
-        public double? Value { get; set; }
+        public float? Value { get; set; }
         /// <summary>
         /// foreign key
         /// </summary>
